refactor(rspulserender): split result exporting and response rendering functionality
response_renderer.h now only handles the rendering of the results. receiver_export now only handles the exporting of the rendered responses to output files.
